fix(types): Various size_t fixes bundled pass III

+++ src/MayaFlux/Buffers/Container/ContainerBuffer.cpp
@@ -106,7 +106,8 @@ void ContainerToBufferAdapter::extract_channel_data(std::span<double> output)
         auto data_span = Kakshya::extract_from_variant<double>(processed_data[0], temp_storage);
 
         auto num_channels = structure.get_channel_count();
-        auto samples_to_copy = std::min(output.size(), data_span.size() / num_channels);
+        auto samples_to_copy = std::min(static_cast<size_t>(output.size()),
+            static_cast<size_t>(data_span.size() / num_channels));
 
         for (auto i : std::views::iota(0UZ, samples_to_copy)) {
             auto interleaved_idx = i * num_channels + m_source_channel;
